+++ src/screens/AllTabs/index.js
@@ -1,0 +1,13 @@
+import React, { Component, type Node } from "react";
+
+type Props = {};
+
+type State = {};
+
+class AllTabs extends Component<Props> {
+  render(): Node {
+    return <div>all tabs</div>;
+  }
+}
+
+export default AllTabs;