+++ src/screens/MobileView/index.js
@@ -1,0 +1,9 @@
+import React, { Component, type Node } from "react";
+
+class MobileView extends Component<{}> {
+  render(): Node {
+    return <div>mobile</div>;
+  }
+}
+
+export default MobileView;