+++ src/screens/NewTab/index.js
@@ -1,0 +1,13 @@
+import React, { Component, type Node } from "react";
+
+type Props = {};
+
+type State = {};
+
+class NewTab extends Component<Props> {
+  render(): Node {
+    return <div>Newtab</div>;
+  }
+}
+
+export default NewTab;