+++ src/screens/TabView/index.js
@@ -4,39 +4,33 @@ import React, { Component, type Node } from "react";
 import { connect } from "react-redux";
 import { bind } from "decko";
 import { type Tab as TabType } from "../../types/tab";
-import { Route } from "react-router-dom";
+import { Route, Switch } from "react-router-dom";
 import UserComponent from "../User";
 import GroupComponent from "../Group/index";
 import TabBar from "../../components/TabBar";
+import EmptyView from "../EmptyView";
+import Box from "../../elements/Box.web";
+import NewTab from "../NewTab";
+import AllTabs from "../AllTabs";
 
-type Props = {
-  tabs: Array<TabType>,
-};
+type Props = {};
 
 class TabView extends Component<Props> {
   render(): Node {
-    const { tabs } = this.props;
     return (
-      <div>
+      <Box full column>
         <TabBar />
 
-        <div>
+        <Switch>
+          <Route exact path="/" component={EmptyView} />
+          <Route exact path="/tabs/new" component={NewTab} />
+          <Route exact path="/tabs/all" component={AllTabs} />
           <Route exact path="/u/:username" component={UserComponent} />
           <Route exact path="/g/:username" component={GroupComponent} />
-        </div>
-      </div>
+        </Switch>
+      </Box>
     );
   }
 }
 
-function mapStateToProps(state, props) {
-  return {
-    tabs: state.tabs.tabs,
-  };
-}
-
-function mapDispatchToProps(dispatch) {
-  return {};
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(TabView);
+export default TabView;